requirements.txt
mysql-connector-python==8.2.0
bcrypt==4.2.1
Pillow==9.5.0
